Rename files to avoid nomenclature clash with n-teaser
