+++ server/transforms/body.js
@@ -5,7 +5,7 @@ const relatedArticleToTeaser = require('./related-article-to-teaser');
 const ariaHiddenMidContentInterruptions = require('./aria-hidden');
 const tableOfContents = require('./table-of-contents');
 const lightSignup = require('./light-sign-up');
-const teaserContent = require('./teaser-content');
+const teaserContent = require('./content-preview');
 const inlineAd = require('./inline-ad');
 const extractToc = require('./extract-toc');
 const extractMainImage = require('./extract-main-image');
