+++ server/transforms/content-preview.js
